+++ ExposurePlayback/Context/Playable/AssetPlayable.swift
@@ -10,22 +10,24 @@ import Foundation
 import Exposure
 
 internal protocol AssetEntitlementProvider {
-    func requestEntitlement(assetId: String, using sessionToken: SessionToken, in environment: Environment, callback: @escaping (PlaybackEntitlement?, ExposureError?) -> Void)
+    func requestEntitlement(assetId: String, using sessionToken: SessionToken, in environment: Environment, callback: @escaping (PlaybackEntitlement?, ExposureError?, HTTPURLResponse?) -> Void)
 }
 
+/// Defines a `Playable` for the specific vod asset
 public struct AssetPlayable: Playable {
+    /// The asset id
     public let assetId: String
     
-    internal var entitlementProvider: AssetEntitlementProvider = AlamofireEntitlementProvider()
+    internal var entitlementProvider: AssetEntitlementProvider = ExposureEntitlementProvider()
     
-    internal struct AlamofireEntitlementProvider: AssetEntitlementProvider {
-        func requestEntitlement(assetId: String, using sessionToken: SessionToken, in environment: Environment, callback: @escaping (PlaybackEntitlement?, ExposureError?) -> Void) {
+    internal struct ExposureEntitlementProvider: AssetEntitlementProvider {
+        func requestEntitlement(assetId: String, using sessionToken: SessionToken, in environment: Environment, callback: @escaping (PlaybackEntitlement?, ExposureError?, HTTPURLResponse?) -> Void) {
             Entitlement(environment: environment,
                         sessionToken: sessionToken)
                 .vod(assetId: assetId)
                 .request()
                 .validate()
-                .response{ callback($0.value, $0.error) }
+                .response{ callback($0.value, $0.error, $0.response) }
         }
     }
 }
@@ -47,12 +49,29 @@ extension AssetPlayable {
     }
     
     internal func prepareAssetSource(environment: Environment, sessionToken: SessionToken, callback: @escaping (ExposureSource?, ExposureError?) -> Void) {
-        entitlementProvider.requestEntitlement(assetId: assetId, using: sessionToken, in: environment) { entitlement, error in
-            if let entitlement = entitlement {
-                callback(AssetSource(entitlement: entitlement, assetId: self.assetId), nil)
+        entitlementProvider.requestEntitlement(assetId: assetId, using: sessionToken, in: environment) { entitlement, error, response in
+            if let value = entitlement {
+                let source = AssetSource(entitlement: value, assetId: self.assetId)
+                source.response = response
+                callback(source, nil)
             }
-            else {
-                callback(nil,error!)
+            else if let error = error {
+                callback(nil,error)
+            }
+        }
+    }
+}
+
+extension AssetPlayable {
+    public func prepareSourceWithResponse(environment: Environment, sessionToken: SessionToken, callback: @escaping (ExposureSource?, ExposureError?, HTTPURLResponse?) -> Void) {
+        entitlementProvider.requestEntitlement(assetId: assetId, using: sessionToken, in: environment) { entitlement, error, response in
+            if let value = entitlement {
+                let source = AssetSource(entitlement: value, assetId: self.assetId)
+                source.response = response
+                callback(source, nil, response)
+            }
+            else if let error = error {
+                callback(nil,error,response)
             }
         }
     }